Remove zero-width spaces.
This appears to do nothing in many diff tools, but it's removing a trailing
zero-width space character from the names of a bunch of method identifiers.
Zero-width spaces are valid identifier characters, so this actually changes
the identifier to what one would expect it should be, and to boot some
compilers can't handle zero-width spaces in identifiers.

Test: make
Bug: 68811146
Change-Id: I3cad25f29a4c011fb62e3462513cf5ce7126081e