Handle lack of error from d2i_X509_bio
BoringSSL changed d2i_X509_bio to not add an entry to the BoringSSL
error queue when given an empty input, which broke a couple tests that
checked that.  Conscrypt doesn't actually care what error is returned,
we just end up throwing CertificateException no matter what happened,
so allow there to be an empty queue and handle it properly like we do
elsewhere.

Test: cts -m CtsLibcoreTestCases -t com.android.org.conscrypt
Test: cts -m CtsLibcoreTestCases -t tests.security.cert
Change-Id: Id21ef45a8a3b416a227210d642e774a8df89ad0d