Annotate repackaged API with @UnsupportedAppUsage
Adds @UnsupportedAppUsage annotations to the parts of conscrypt that
are in the hidden API. Allows the conscrypt entries to be removed from
the frameworks/base/config/hiddenapi-light-greylist.txt.

This was tested by making and then manually checking that the generated
out/target/common/obj/PACKAGING/hiddenapi-light-greylist.txt was the
same (after sorting) before and after this change.

Bug: 117818301
Test: see above
Change-Id: I3820dfb7921954b93b81da3e9f355c2b807130b9