Add missing default constructors
Some IntraCoreApi annotations were missing as they were targeted at
default constructors and so were not present in the source. This change
adds the default constructors which allows the missing annotations to
be added.

Also, Metalava was accidentally including default constructors in the
API generated for CorePlatformApi and IntraCoreApi annotations. This
change adds the constructor and appropriate annotations to allow the
metalava bug to be fixed without changing the API.

Tested by building the API with and without the Metalava fix and
checking that they were the same.

Bug: 117818301
Bug: 119859108
Test: make checkbuild
Change-Id: I3fac495a784efb68ff117a1d0a36f0b799cdc49b