Switch from block comments to line comments
The code that parses this is too lenient, it allows too many types of
comments which makes it difficult to process reliably. This change will
allow it to be changed to only allow line comments which are simpler to
process.

Tested by running the srcgen/generate_android_src.sh script and
verifying that it did not change the generated code.

Bug: 117818301
Test: see above
Change-Id: Ied5028d594352939b7df068b888b84d750868a19