Revert "Merge upstream master"
This reverts commit 2136bd7e5eda8f3bb412001b974c4016f3837d8a.

Reason for revert: Broke com.squareup.okhttp.CallTest#cancelDuringHttpsConnect
(but not com.squareup.okhttp.CallTest#cancelDuringHttpConnect).

My theory is that the change caused threads waiting on a secure
socket to not wake up when the socket was closed.
ServerSocketConcurrentCloseTest only uses plain (non-TLS) sockets
so would not have noticed this. It might be worth expanding it
to cover secure sockets in a follow-up CL.

Bug: 123121395
Bug: 122881655
Test: make libconscrypt_jni
Test: Checked that the following fails before but passes after this CL:
      atest CtsLibcoreOkHttpTestCases:com.squareup.okhttp.CallTest#cancelDuringHttpsConnect

Change-Id: I355130ad4d1c835e70839b36f718b915518015a7