boringssl_self_test{32,64}: Copy to /system/bin.
This is the second attempt to submit this CL. The first attempt
accidentally droped the copy from /apex. This CL omits the change
to apex/Android.bp

/system and /apex/com.android.conscrypt each carry a copy of libcrypto.
Therefore the boringssl_self_test{32,64} binaries are likewise needed
in both /system/bin/ and /apex/com.android.conscrypt/bin/. Before this
CL, they existed in the latter. The effect of this CL is to also make
them available in the former.

Because borignssl_self_test is no longer specific to apex/, it also
makes sense to move the selftext/ folder with the cc_binary build
target definition and sources out of the apex/ folder, but this is
left to a future CL.

Test: Ran the following command and checked that each of the four
      files exists on a device with 64+32bit ABIs:
      adb shell su 0 ls -l \
      /apex/com.android.conscrypt/bin/boringssl_self_test32 \
      /apex/com.android.conscrypt/bin/boringssl_self_test64 \
      /system/bin/boringssl_self_test32 \
      /system/bin/boringssl_self_test64
Test: Ran the corresponding command with "file" in place of
      "ls -l" and confirmed manually that each of those
      binaries is for the expected (32 vs. 64bit) ABI.
Bug: 137267623

Change-Id: Ic394817ea2f882494ffdc1eea7fdc2746adf4f66