Add no-op ConscryptEngineSocket.setHandshakeTimeout.
TLS handshake timeout is not settable via the JSSE API and this class
does not implement it.  However ConscryptFileDescriptorSocket provides
a setter which is used by SSLCertificateSocketFactory and a very small
number of apps, and the default implementation in AbstractConscryptSocket
throws which causes compat issues when we switch to ConscryptEngineSocket.

Having a no-op implementation is a behaviour change for
SSLCertificateSocketFactory but that class is deprecated.  We can add
a concrete implementation if it proves problematic.

Needs upstreaming to Github.

Bug: 146041327
Test: atest SSLCertificateSocketFactoryTest
Test: atest CtsNetSecConfigBasicDebugEnabledTestCases
Test: atest CtsNetSecConfigBasicDomainConfigTestCases
Test: atest CtsNetSecConfigCleartextTrafficTestCases
Test: atest CtsNetSecConfigInvalidPinTestCases
Test: atest CtsNetSecConfigPrePCleartextTrafficTestCases
Change-Id: I106cd35748bd890e5821afd33a3927e556e367e6