Temporarily comment out conscrypt-module-host-exports
The sdk/module_exports do not support building a single snapshot for
multiple os types. Instead it builds a separate snapshot per os type
which would conflict when unpacked. Attempting to fix that issue is
complicated by the fact that there are presubmit builds which rely
on that behavior for handling the conscrypt-module-host-exports.

The quickest and safest way to deal with that is to temporarily
comment out the broken module, fix the build issues and then
uncomment it.

Bug: 150451422
Test: m nothing
Change-Id: I2f75255b20c3614de5f10c266f9d5aca7574f4b7