Mark the entire conscrypt core/platform API as stable.
It already is de facto stable, because conscrypt was updatable in
Android 10. To make this situation clear to everyone, the annotations
should reflect that. (The .java changes were generated using new
functionality in currysrc, which is triggered by renaming the
file under srcgen.)

This is a no-op as far as the actual API surfaces is concerned.

Test: m checkapi
Test: m conscrypt
Bug: 157640067
Change-Id: Id9da8dd6537c94e139b13bc1a44003129e46b031