Fix ConscryptEngineSocket.close().
Previously, this method closed the underlying socket first and then
the SSLEngine.  However closing a connected SSLEngine queues a TLS
close notification which obviously can't be sent if the socket is
closed.  Also, the pending bytes prevent the engine from freeing
its native resources including pipe file descriptors until the
SSLEngine is eventually garbage collected.

This change is split out from previous Conscrypt upstream merge
as it was generating exceptions on TreeHugger cf_x86 which were
not reproducible on acloud cf_x86 or real devices. As a temporary
workaround to get the TLS close fix in this change landed, the
relevant test ignores that exception.

Bug: 159199048
Test: atest CtsLibcoreTestCases
Change-Id: I416d00dc135b790a2993e0fed45fdee78284c04f