Add native coverage flushing to ConscryptInstrumentationListener.
Temporary (I hope) home for this until the coverage
infrastructure catches up.

Adds a testRunFinished() method to
ConscryptInstrumentationListener which will detect if the
signal handler for flushing native coverage is installed and
if so invoke it and wait for it to complete.

Bug: 160623274
Test: atest MtsLibcoreTestCases (both instrumented for coverage, and not).
Change-Id: I68ddced880d79fdc5b0252e6e18f0b2c5686de40
(cherry picked from commit a8d751d2fdaaeaf3b3fa8c582effae39d5009073)