conscrypt: mark libjavacrypto as jni_libs
This helps the build system to convey the information to linkerconfig so
that conscrypt linker namespace is visible and libnativeloader can load
the lib.

Compatibility:
The file entries of the apex doesn't change except apex_manifest.pb
which has an additional key. ("jniLibs: ["libjavacrypto.so"]). So this
doesn't affect Q. However, linkerconfig in R would try to make conscrypt
apex visible because it has jni libs now. It is no-op because
linkerconfig in R already makes conscrypt apex visible, which is
hard-coded.

Bug: 150767721
Test: presubmit
Change-Id: I526c9269128c5fceaa87ecc9af5ea58a186731e4