Copy public API artifacts to dist as conscrypt...
Uses the new dist_stem property to specify the base name of the
artifacts to be conscrypt to match the names of the files in
prebuilts/sdk/30. That property also affects the filegroup name that
references the the latest released versions of the API and removed API
specification files against which the latest API is checked for
compatibility.

Removes the no_dist: true property that would otherwise prevent the
artifacts being copied to the dist and removes the filegroups that
bridged the inconsistency between the names generated by prebuilt_apis
in prebuilts/sdk and the names used by the conscrypt.module.public.api
module. Those filegroups are no longer needed because setting the
dist_stem: "conscrypt" causes conscrypt.module.public.api to use the
same names as generated by prebuilt_apis.

Bug: 173715943
Test: lunch sdk-eng && m dist sdk
      Run above before and after this change to ensure that the
      conscrypt stubs jar and API specification file (but not the
      removed API specification file) are copied to the dist.
      Also, added a new method to the
      prebuilts/sdk/30/public/api/conscrypt.txt and run m checkapi
      to ensure that incompatible changes between the latest released
      API and the current API are detected properly.
Change-Id: I00b55534472378e2e62eb7c0bc37c2e3a1b1d3d9