[DO NOT MERGE] CTS 11: Add key size for XDH.
Part of a set of changes to Android 11 CTS to enable it to pass
with or without an Android 12-based Conscrypt module installed.

XDH KeyFactory may or may not be present depending on which
Conscrypt module is installed. Provide a key size for it
so that if it is present it can be tested.

Bug: 195615712
Test: atest CtsLibcoreTestCases with and without a 12-based Conscrypt module
Change-Id: I72bce2d4f3be1bf1410b622252ef10c10f2e7f75