[DO NOT MERGE] CTS 11: Work around blocklist class name change.
Part of a set of changes to Android 11 CTS to enable it to pass
with or without an Android 12-based Conscrypt module installed.

In the Android 12 based Conscrypt module, the certificate blocklist
class was renamed, but the old class name is used in the Android 11 CTS
suite.

As this only affects a single test and as this test is exercised by
Conscrypt MTS before the module can be pushed, simply skip the test if
the expected class is unavailable.

On an Android 11 device with an earlier Conscrypt module, the test
will run as before.

Required upgrading the test to Junit 4 to allow assumptions to work
correctly.

Bug: 195615983
Test: atest CtsLibcoreTestCases with and without a 12-based Conscrypt module
Change-Id: Ibffb3f263af9c68804bc2e770f35617beb7d734d