[DO NOT MERGE] CTS 10: Remove invalid count check of unwrap calls.
Part of a set of changes to Android 10 CTS to enable it to pass
with or without an Android 12-based Conscrypt module installed.

This change is a backport of the equivalent CTS 11 change
in https://r.android.com/1833496

Android 10 CTS tries to enforce buggy SSLEngine.wrap() behaviour
where it expects multiple buffers passed to wrap() to generate
multiple TLS records.  This behaviour is fixed in the Android 12
based Conscrypt module where multiple buffers are coalesced into
a single TLS record where possible.

Simply removing the check from Android 10 CTS as it has no value.

Bug: 197500794
Bug: 195615524
Test: atest CtsLibcoreTestCases with and without a 12-based Conscrypt
module

Change-Id: I6f276ede6029c234c5600d41778ab00e78a4b40a