[DO NOT MERGE] CTS 10: Add key size for XDH.
Part of a set of changes to Android 10 CTS to enable it to pass
with or without an Android 12-based Conscrypt module installed.

This change is a backport of the equivalent CTS 11 change
in https://r.android.com/1833497

XDH KeyFactory may or may not be present depending on which
Conscrypt module is installed. Provide a key size for it
so that if it is present it can be tested.

Bug: 197500794
Bug: 195615712
Test: atest CtsLibcoreTestCases with and without a 12-based Conscrypt
module

Change-Id: Id3b1aa1a40b31e6011f04a8886f4e31fdc2ffc83