Keep Conscrypt certs in internal up to date with system/
Cherry-pick note: Root CA certificates are updated in internal-master
for historical reasons.  However in UDC we are also shipping an
updatable copy of them in the Conscrypt Mainline module, which is
AOSP-first and used in preference to the certificates in /system on UDC
and later.

So, without this change, the Conscrypt module is downgrading the system
certificates to what we shipped with T, because that's what's currently
in AOSP.

And unfortunately, we never shipped a root certificate update with T so
they are downgraded to what we shipped in S.

Test: atest CertificateTest
Bug: 265189774
Change-Id: I6a13dddcf80ee7aefbf3777ca775ebb8c9909f8e
Merged-In: I6a13dddcf80ee7aefbf3777ca775ebb8c9909f8e
(cherry picked from commit 802d82931a1a08230cc327010695f187ffc449d7)