Validate log states when loading the log list
The log list contains timestamps describing when a log entered its
state. In theory, this time should be in the past as only useful log
details are sent to the device. This assumption may break if the
device's clock is incorrect; or if the server pushes a list with
timestamps in the future. If that was to happen, the SCT validation
would fail because the log would be in an STATE_UNKNOWN.

Avoid these scenarios by validating the state using the current time. If
an unknown state is found, fail the loading of the log list and
fail-open the CT verification.

Bug: 409857378
Test: atest ConscryptPrivateTestCases
Flag: EXEMPT bugfix
Change-Id: I4e171eed746fc2f0ba9f51a73c4126e9f4e7d1cf