Move TimeSupplier to System.currentTimeMillis
In commit ebe9322c379 ("Add periodic reloading for CT log list"),
TimeSupplier was added to LogStoreImpl to help with the reloading of the
log list file. This supplier was based on System.nanoTime() which
returns the time since the start of the process. This is adequate to
check if the log list file should be reloaded.

In commit 9eb41ea11ae ("Validate log states when loading the log list"),
TimeSupplier was reused to validate the state of the log when parsing
the log list. This is an issue because:
  1. TimeSupplier relied on nanoseconds and not milliseconds;
  2. TimeSupplier relied on the start time, not the wall time.

Move TimeSupplier to use the wall time, in milliseconds. Update the
tests accordingly.

The process time was previously picked to guarantee that no time travel
in the past would be observed. Because we are moving to the wall time,
support the case where the device clock is moved backwards.

Bug: 414690762
Test: atest ConscryptPrivateTestCases
Test: atest NetSecConfigCertificateTransparencySctLogListTestCases
Flag: EXEMPT bugfix
Change-Id: I86538c6f699998a076d47f78770f0c884cd0487b