Add NSC classes to framework-conscrypt-nsc
1. These are the classes from the frameworks NSC code.

2. The only difference is that I removed the check for the CT aconfig
flag from the ApplicationConfig class. I was having troubles adding the
flag dependency to the package, and this implementation will never see
the flag value as false, so I thought it would be easier to just remove
the flag.

3. I modified the namespace to make sure collisions are avoided between
the old and new implementation. Because of the requirements of aconfig
and build flags, there will be a period of time when both
implementations co-exist, until the necessary flags are all ramped up
and it will be possible to remove them.

4. The existing code already has unit as well as CTS test coverage. We
will not introduce new tests, but as part of the migration we will
migrate the existing ones.

Bug: 397646538
Test: build-flag --all-releases set RELEASE_CONSCRYPT_NSC true && m
Flag: build.release_conscrypt_nsc
Change-Id: I7429fa4bc5b7464c498559706cbc82ee1e013e47