Report hit on blocklist
Define a new CertBlocklistImpl.Builder class to better control the
instantiation of CertBlocklistImpl.

The internal Sets of CertBlocklistImpl are refactored to Maps, in order
to return the associated blocklist entry.

The cache values are updated to Optional<Entry> to either capture an
entry in the blocklist; or if no entry was previously found, the empty
value.

Bug: 425499049
Test: atest ConscryptPrivateTestCases
Test: atest CtsSecurityTestCases:android.security.cts.CertBlocklistTest
Test: atest CtsSecurityTestCases:android.security.cts.CertBlocklistFileTest
Flag: EXEMPT bugfix
Change-Id: I6430a86a3812f80d73d56269229460cf85c6f6c0