Move Conscrypt NSConfig classes to android.security package
According to the Mainline team, if we are not interested in backporting
the NSConfig to pre-17 Android, we should be able to migrate the code to
Mainline without renaming the package, as long as the java library has
37 in the min_sdk_version field.

This would be an advantage as it would mean we do not need to deprecate
the NSPolicy public API, we could eventually just replace it with the
Mainline implementation.

This CL also brings the Conscrypt nsc classes up to date with all the
recent changes that were applied to the nsc classes in frameworks.

Changes included:
  Ibec4a3e3bc61fcddf0bdf9634d90372633dcf0a2
  Ie3feca711d5e5f90e0cbacfe0efd8d45e9864322
  I6c652607b7fadc314cca7d169d8a84c9ed8e3078
  Iaf0d6f943c83f7e2c370d4d2136d2ad6cb659dd3
  I4681820ae88df1f17cfc4bfe4bdc2ad94138f696
  Ia2384411345409b2ffc7af59017aca1423014d86

Bug: 397646538
Test: m framework-conscrypt-nsc
Flag: build.release_conscrypt_nsc
Flag: android.security.conscrypt_network_security_config
Flag: com.android.org.conscrypt.net.flags.network_security_config
Change-Id: I49ed8bc54ea31a9a18b868b1ce042fddc2e81f3a