Add ConscryptX509TrustManager interface
The TrustManagerImpl class implements methods that go beyond the
standard X509TrustManager or X509ExtendedTrustManager classes. The
Android framework expects these methods to be available. Access to these
methods is usually done by duck-typing which reduces the ability to
catch issues at compilation time.

Formalize these extra methods with a new interface
ConscryptX509TrustManager.

On Android, note that these methods are not directly accessible to apps.
The intermediate x509TrustManagerExtensions class should also be updated
for that purpose.

The intent is to replace any direct reference to TrustManagerImpl in the
Android framework with this interface.

This change is binary compatible.

FIXED_READ_ONLY_FLAG_REASON=MAINLINE_BETA_NATIVE_FLAG

Bug: 450136572
Bug: 448210434
Test: presubmit
API-Coverage-Bug: 449878498
Change-Id: Ib837d6a61788eef2573c7926808df871c0932ae9