Add API to attach a NetworkSecurityPolicy to a TrustManager
TrustManagerImpl is implemented for all of Conscrypt's backends,
including OpenJDK. Therefore, it cannot rely on Android-only classes,
such as libcore.net.NetworkSecurityPolicy.

Introduce a new class ConscryptNetworkSecurityPolicy for each backend.
For the platform backend (mainline), the constructor accepts a
libcore.net.NetworkSecurityPolicy to which it delegates its logic.

Internally, the ConscryptNetworkSecurityPolicy implements
NetworkSecurityPolicy which contains the methods expected by Conscrypt
components (e.g., TrustManagerImpl or SSLParametersImpl). This is to
ensure that all backend implementations remain in sync.

The previous Java property logic is removed for the openjdk and android
backends as no implementation was ever provided. If required in the
future, it can be brought back.

Bug: 448210434
Test: atest --include-subdirs cts/tests/tests/networksecurityconfig --all
API-Coverage-Bug: 449878498
Change-Id: I5fcc58a797a2862c286b6961bbc401007f95b6b2