Fix the target SDK gate for ECH mainline Conscrypt NSC and add app compat unit tests
I mistakenly thought the @EnabledAfter tag was inclusive, and did not realize it was exclusive. ECH should be enabled from 37+, which means it should actually be BAKLAVA marked as the last targetSdkVersion that is unsupported. Also add unit tests to test for the change in default mode based on CompatChanges.

Bug: 450269491
Flag: com.android.org.conscrypt.net.encrypted_client_hello_platform
Test: atest ConscryptNetworkSecurityConfigTests
Change-Id: I4fcb4d07e29ee108cc76df5e695703e834770030