Fallback if CT verification reason is unknown
On Baklava, getCertificateTransparencyVerificationReason is not
overridden by the Platform ConfigNetworkSecurityPolicy. In this case,
the placeholder value from the libcore default NetworkSecurityPolicy is
returned, UNKNOWN.

Instead of returning this value, fallback to the previous method of
inferring the reason, based on the behaviour of
isCertificateTransparencyVerificationRequired.

Bug: 425786616
Test: presubmit
Flag: com.android.libcore.network_security_policy_reason_ct_enabled_api
Change-Id: I414a5277c384788fd0dd15c87e148bd3c595747a