Stop ignoring conscrypt's previously released platform API
Previously, `conscrypt.module.platform.api` set
`unsafe_ignore_missing_latest_api: true` which meant that it would
ignore the previously released API (in
`prebuilts/sdk/36.1/public/api/conscrypt-coreplatform.txt`) and instead
use a blank file.

That meant it would not handle using `@FlaggedApi` to make changes in
an existing API. That is because if the flag was disabled it would try
and revert it to the previous version of the API but there was no
previous version of the API so it would assume it was new and remove
it.

This change removes the `unsafe_ignore_missing_latest_api: true`
setting and the `api/platform/last-*` files which were used before
`unsafe_ignore_missing_latest_api: true` was set in ag/14841303 but
are now out of date and unused.

That revealed quite a few lint errors which were previously suppressed
by the `unsafe_ignore_missing_latest_api: true` setting so they are as
baselined as part of this change.

Bug: 460287792
Test: m checkapi
Change-Id: I7357c194b706645a44dd0c2cc5838720a09c91c8