Retry: Stop ignoring conscrypt's previously released platform API
Previously, `conscrypt.module.platform.api` set
`unsafe_ignore_missing_latest_api: true` which meant that it would
ignore the previously released API (in
`prebuilts/sdk/36.1/public/api/conscrypt-coreplatform.txt`) and instead
use a blank file.

That meant it would not handle using `@FlaggedApi` to make changes in
an existing API. That is because if the flag was disabled it would try
and revert it to the previous version of the API but there was no
previous version of the API so it would assume it was new and remove
it.

This change removes the `unsafe_ignore_missing_latest_api: true`
setting.

That revealed quite a few lint errors which were previously suppressed
by the `unsafe_ignore_missing_latest_api: true` setting so they are
baselined as part of this change.

The previous attempt to submit this failed in `next` due to a bug in
Metalava. That bug is fixed in https://r.android.com/3865474 and is
being merged into Android in http://ag/37159221.

Bug: 460287792
Test: m checkapi
Change-Id: Ice47688e74a405375ae7451b3d9d9d7e8659c412