Dist libconscrypt_openjdk_jni for sdk-repo-build-tools
The prebuilt signapk.jar is stored in prebuilts/sdk/tools/lib,
but it is not distributed as part of the SDK so it has historically
been manually updated.  Dist its JNI libraries as part of
sdk-repo-build-tools so that the prebuilts update script can update
it alongside the rest of the official SDK tools.

Bug: 325130426
Test: manual
Change-Id: Ie3182876d62e62426140ea333e9ce3f4a76de1d8