Enable DEPRECATE_USES_CLEARTEXT_TRAFFIC compat change after Cinnamon Bun.
This was always the plan, but I thought @EnableAfter was inclusive, and that I had to wait for the new version to be in the Enum.
It's not, so this change will be enabled for the release following 26Q2 as it should be.

Bug: 415007211
Flag: android.security.deprecate_uses_cleartext_traffic2
Change-Id: I688fe481c38a653dac7a9f91317b5c01469917ce