Add support for flatbuffers log list
The log list format is changed from JSON to flatbuffers. This allows
Conscrypt to map the file in memory without further parsing. When a log
is requested, the list is queried using a binary search. LogInfo are
then constructed on demand and kept in a cache.

These changes are implemented in a separate log store: LogStoreImplv3.
The existing implementation is moved to LogStoreImplv2. The flag
ct_compat_version_3 is used to decide if v3 is used.

Bug: 407952621
Test: presubmit
Flag: com.android.org.conscrypt.net.flags.ct_compat_version_3
Change-Id: Icb49904fe25eb08029d381d3a68ac0a3536f86e7