Add LogStore.InvalidLogException
When using the flatbuffers representation, we may become aware of a log
list corruption while verifying an SCT (e.g., invalid public key). In
this case, we need to surface this information so that the verification
is non-blocking (i.e., fail-safe for corrupted log list). Add a new
exception which can be raised within LogStore.getKnownLog for that
purpose.

Test: presubmit
Bug: 456392970
Flag: com.android.org.conscrypt.net.flags.ct_compat_version_3
Change-Id: Idcc702572c320b6e70f690d06836c2adb0d9ab78